+++ Makefile.common
@@ -716,7 +716,12 @@ ifeq ($(HAVE_XMB), 1)
 endif
 
 ifeq ($(HAVE_OZONE), 1)
-   OBJ += menu/drivers/ozone.o
+   OBJ += menu/drivers/ozone/ozone.o
+   OBJ += menu/drivers/ozone/ozone_entries.o
+   OBJ += menu/drivers/ozone/ozone_display.o
+   OBJ += menu/drivers/ozone/ozone_texture.o
+   OBJ += menu/drivers/ozone/ozone_theme.o
+   OBJ += menu/drivers/ozone/ozone_sidebar.o
    DEFINES += -DHAVE_OZONE
    HAVE_MENU_COMMON = 1
 endif
